refactord code
